+++ CHANGELOG.md
@@ -455,3 +455,4 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 * Make Spriter animation public (kiooeht)
 * Spriter animations for monsters (kiooeht)
 * Multi-word keywords (kiooeht)
+* What mod is this from? for events (kiooeht)